Change made to data type upon request

+++ AddressesAPI/Infrastructure/Migrations/AddressesContextModelSnapshot.cs
@@ -195,7 +195,7 @@ namespace AddressesAPI.Infrastructure.Migrations
 
                     b.Property<int?>("UnitNumber")
                         .HasColumnName("unit_number")
-                        .HasColumnType("integer");
+                        .HasColumnType("smallint");
 
                     b.Property<string>("UsageCode")
                         .HasColumnName("blpu_class")
@@ -363,7 +363,7 @@ namespace AddressesAPI.Infrastructure.Migrations
 
                     b.Property<int?>("UnitNumber")
                         .HasColumnName("unit_number")
-                        .HasColumnType("integer");
+                        .HasColumnType("smallint");
 
                     b.Property<string>("UsageCode")
                         .HasColumnName("blpu_class")
